+++ cookbooks/aws-parallelcluster-platform/resources/nvidia_nvlsm/nvidia_nvlsm_amazon2.rb
@@ -1,0 +1,27 @@
+# frozen_string_literal: true
+
+# Copyright:: 2025 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License").
+# You may not use this file except in compliance with the License.
+# A copy of the License is located at
+#
+# http://aws.amazon.com/apache2.0/
+#
+# or in the "LICENSE.txt" file accompanying this file.
+# This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, express or implied.
+# See the License for the specific language governing permissions and limitations under the License.
+
+provides :nvidia_nvlsm, platform: 'amazon', platform_version: '2'
+
+use 'partial/_nvidia_nvlsm_common.rb'
+use 'partial/_nvidia_nvlsm_rhel.rb'
+
+def nvlsm_installation_enabled?
+  # NVLSM is not supported on Amazon Linux 2.
+  false
+end
+
+def platform
+  "amzn#{node['platform_version'].to_i}"
+end